requirement.txt
autopep8==1.3.3
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
Django==1.11.6
djangorestframework==3.7.0
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
pycodestyle==2.3.1
Pygments==2.2.0
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
